+++ lib/rails/resource_service.rb
@@ -11,7 +11,13 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         index_resource_url ||= "/#{plural_resource_name}.json"
-        HTTP.get(index_resource_url, payload: index_show_destroy_resource_params(params: params.to_h), &response_handler)
+        HTTP.get(index_resource_url, payload: index_show_destroy_resource_params(params: params.to_h)) do |response|
+          if response.ok? && !resource_class.nil?
+            resource_response_objects = Native(response.body)
+            resources = resource_response_objects.map { |resource_response_object| build_resource_from_response_object(resource_class:, resource_response_object:) }
+          end
+          response_handler.call(response, resources)
+        end
       end
       
       def show(resource: nil, resource_class: nil, resource_id: nil, singular_resource_name: nil, plural_resource_name: nil, show_resource_url: nil, params: nil, &response_handler)
@@ -20,7 +26,13 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         show_resource_url ||= "/#{plural_resource_name}/#{resource_id}.json"
-        HTTP.get(show_resource_url, payload: index_show_destroy_resource_params(params: params.to_h), &response_handler)
+        HTTP.get(show_resource_url, payload: index_show_destroy_resource_params(params: params.to_h)) do |response|
+          if response.ok? && !resource_class.nil?
+            resource_response_object = Native(response.body)
+            resource = build_resource_from_response_object(resource_class:, resource_response_object:)
+          end
+          response_handler.call(response, resource)
+        end
       end
       
       def create(resource: nil, resource_class: nil, resource_attributes: nil, singular_resource_name: nil, plural_resource_name: nil, create_resource_url: nil, params: nil, &response_handler)
@@ -28,7 +40,13 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         create_resource_url ||= "/#{plural_resource_name}.json"
-        HTTP.post(create_resource_url, payload: create_update_resource_params(resource:, resource_class:, resource_attributes:, singular_resource_name:, params: params.to_h), &response_handler)
+        HTTP.post(create_resource_url, payload: create_update_resource_params(resource:, resource_class:, resource_attributes:, singular_resource_name:, params: params.to_h)) do |response|
+          if response.ok? && !resource_class.nil?
+            resource_response_object = Native(response.body)
+            resource = build_resource_from_response_object(resource_class:, resource_response_object:)
+          end
+          response_handler.call(response, resource)
+        end
       end
       
       def update(resource: nil, resource_class: nil, resource_id: nil, resource_attributes: nil, singular_resource_name: nil, plural_resource_name: nil, update_resource_url: nil, params: nil, &response_handler)
@@ -37,7 +55,13 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         update_resource_url ||= "/#{plural_resource_name}/#{resource_id}.json"
-        HTTP.patch(update_resource_url, payload: create_update_resource_params(resource:, resource_class:, resource_attributes:, singular_resource_name:, params: params.to_h), &response_handler)
+        HTTP.patch(update_resource_url, payload: create_update_resource_params(resource:, resource_class:, resource_attributes:, singular_resource_name:, params: params.to_h)) do |response|
+          if response.ok? && !resource_class.nil?
+            resource_response_object = Native(response.body)
+            resource = build_resource_from_response_object(resource_class:, resource_response_object:)
+          end
+          response_handler.call(response, resource)
+        end
       end
       
       def destroy(resource: nil, resource_class: nil, resource_id: nil, singular_resource_name: nil, plural_resource_name: nil, destroy_resource_url: nil, params: nil, &response_handler)
@@ -70,6 +94,12 @@ module Rails
       def singular_resource_name_for_resource_class(resource_class)
         return nil if resource_class.nil?
         resource_class.to_s.split('::').last.downcase
+      end
+      
+      def build_resource_from_response_object(resource_class:, resource_response_object:)
+        resource = resource_class.new
+        resource_response_object.each { |attribute, value| resource.send("#{attribute}=", value) rescue nil }
+        resource
       end
     end
   end